Permite crear un viaje que a su vez, crea una respectiva quota de viajes y quotas y dicha cuota con su hotel.

+++ db/migrate/20130328223940_create_assets.rb
@@ -1,6 +1,6 @@
 class CreateAssets < ActiveRecord::Migration
   def change
-    drop_table :assets
+    drop_table :assets if self.table_exists?("assets")
     create_table :assets do |t|
       t.string :name
       t.references :hotel
